+++ advanced/kernel/src/main/java/org/neo4j/kernel/impl/persistence/EntityIdGenerator.java
@@ -21,7 +21,7 @@ package org.neo4j.kernel.impl.persistence;
 
 public interface EntityIdGenerator
 {
-    int nextId( Class<?> clazz );
+    long nextId( Class<?> clazz );
 
     long getHighestPossibleIdInUse( Class<?> clazz );
